Windows: Crashlog: Fix printf format warnings

+++ src/os/windows/crashlog_win.cpp
@@ -163,27 +163,27 @@ static const char *GetAccessViolationTypeString(uint type)
 	if (message != nullptr && strcasestr(message, "out of memory") != nullptr) {
 		PROCESS_MEMORY_COUNTERS pmc;
 		if (GetProcessMemoryInfo(GetCurrentProcess(), &pmc, sizeof(pmc))) {
-			buffer += seprintf(buffer, last, " WorkingSetSize: " PRINTF_SIZE "\n", pmc.WorkingSetSize);
-			buffer += seprintf(buffer, last, " PeakWorkingSetSize: " PRINTF_SIZE "\n", pmc.PeakWorkingSetSize);
-			buffer += seprintf(buffer, last, " QuotaPagedPoolUsage: " PRINTF_SIZE "\n", pmc.QuotaPagedPoolUsage);
-			buffer += seprintf(buffer, last, " QuotaPeakPagedPoolUsage: " PRINTF_SIZE "\n", pmc.QuotaPeakPagedPoolUsage);
-			buffer += seprintf(buffer, last, " QuotaNonPagedPoolUsage: " PRINTF_SIZE "\n", pmc.QuotaNonPagedPoolUsage);
-			buffer += seprintf(buffer, last, " QuotaPeakNonPagedPoolUsage: " PRINTF_SIZE "\n", pmc.QuotaPeakNonPagedPoolUsage);
-			buffer += seprintf(buffer, last, " PagefileUsage: " PRINTF_SIZE "\n", pmc.PagefileUsage);
-			buffer += seprintf(buffer, last, " PeakPagefileUsage: " PRINTF_SIZE "\n\n", pmc.PeakPagefileUsage);
+			buffer += seprintf(buffer, last, " WorkingSetSize: " PRINTF_SIZE "\n", (size_t)pmc.WorkingSetSize);
+			buffer += seprintf(buffer, last, " PeakWorkingSetSize: " PRINTF_SIZE "\n", (size_t)pmc.PeakWorkingSetSize);
+			buffer += seprintf(buffer, last, " QuotaPagedPoolUsage: " PRINTF_SIZE "\n", (size_t)pmc.QuotaPagedPoolUsage);
+			buffer += seprintf(buffer, last, " QuotaPeakPagedPoolUsage: " PRINTF_SIZE "\n", (size_t)pmc.QuotaPeakPagedPoolUsage);
+			buffer += seprintf(buffer, last, " QuotaNonPagedPoolUsage: " PRINTF_SIZE "\n", (size_t)pmc.QuotaNonPagedPoolUsage);
+			buffer += seprintf(buffer, last, " QuotaPeakNonPagedPoolUsage: " PRINTF_SIZE "\n", (size_t)pmc.QuotaPeakNonPagedPoolUsage);
+			buffer += seprintf(buffer, last, " PagefileUsage: " PRINTF_SIZE "\n", (size_t)pmc.PagefileUsage);
+			buffer += seprintf(buffer, last, " PeakPagefileUsage: " PRINTF_SIZE "\n\n", (size_t)pmc.PeakPagefileUsage);
 		}
 		PERFORMANCE_INFORMATION perf;
 		if (GetPerformanceInfo(&perf, sizeof(perf))) {
-			buffer += seprintf(buffer, last, " CommitTotal: " PRINTF_SIZE "\n", perf.CommitTotal);
-			buffer += seprintf(buffer, last, " CommitLimit: " PRINTF_SIZE "\n", perf.CommitLimit);
-			buffer += seprintf(buffer, last, " CommitPeak: " PRINTF_SIZE "\n", perf.CommitPeak);
-			buffer += seprintf(buffer, last, " PhysicalTotal: " PRINTF_SIZE "\n", perf.PhysicalTotal);
-			buffer += seprintf(buffer, last, " PhysicalAvailable: " PRINTF_SIZE "\n", perf.PhysicalAvailable);
-			buffer += seprintf(buffer, last, " SystemCache: " PRINTF_SIZE "\n", perf.SystemCache);
-			buffer += seprintf(buffer, last, " KernelTotal: " PRINTF_SIZE "\n", perf.KernelTotal);
-			buffer += seprintf(buffer, last, " KernelPaged: " PRINTF_SIZE "\n", perf.KernelPaged);
-			buffer += seprintf(buffer, last, " KernelNonpaged: " PRINTF_SIZE "\n", perf.KernelNonpaged);
-			buffer += seprintf(buffer, last, " PageSize: " PRINTF_SIZE "\n", perf.PageSize);
+			buffer += seprintf(buffer, last, " CommitTotal: " PRINTF_SIZE "\n", (size_t)perf.CommitTotal);
+			buffer += seprintf(buffer, last, " CommitLimit: " PRINTF_SIZE "\n", (size_t)perf.CommitLimit);
+			buffer += seprintf(buffer, last, " CommitPeak: " PRINTF_SIZE "\n", (size_t)perf.CommitPeak);
+			buffer += seprintf(buffer, last, " PhysicalTotal: " PRINTF_SIZE "\n", (size_t)perf.PhysicalTotal);
+			buffer += seprintf(buffer, last, " PhysicalAvailable: " PRINTF_SIZE "\n", (size_t)perf.PhysicalAvailable);
+			buffer += seprintf(buffer, last, " SystemCache: " PRINTF_SIZE "\n", (size_t)perf.SystemCache);
+			buffer += seprintf(buffer, last, " KernelTotal: " PRINTF_SIZE "\n", (size_t)perf.KernelTotal);
+			buffer += seprintf(buffer, last, " KernelPaged: " PRINTF_SIZE "\n", (size_t)perf.KernelPaged);
+			buffer += seprintf(buffer, last, " KernelNonpaged: " PRINTF_SIZE "\n", (size_t)perf.KernelNonpaged);
+			buffer += seprintf(buffer, last, " PageSize: " PRINTF_SIZE "\n", (size_t)perf.PageSize);
 			buffer += seprintf(buffer, last, " HandleCount: %u\n", (uint)perf.HandleCount);
 			buffer += seprintf(buffer, last, " ProcessCount: %u\n", (uint)perf.ProcessCount);
 			buffer += seprintf(buffer, last, " ThreadCount: %u\n\n", (uint)perf.ThreadCount);
